修复接口数据类型错误问题

+++ app/src/main/java/com/goldze/mvvmhabit/aioui/bean/MusicDetailRequest.kt
@@ -12,7 +12,7 @@ data class MusicDetailRequest(
 )
 
 class MusicDetailRequestBody {
-    var deptId: Int = 0
+    var deptId: Long = 0
     var id: Long = 0
     var isDel: Int = 0
     var pageNum: Int = 0
